Create CODE_OF_CONDUCT.md
A code of conduct helps set ground rules for the behavior of your project’s participants. It defines standards for how to engage in a community.

GitHub provides templates for common codes of conduct to help you quickly add one to your project. To add a code of conduct to your project, complete the following steps:

Add a new file named CODE_OF_CONDUCT.md to the root folder of the repository. The Choose a code of conduct template button is displayed.

Click the Choose a code of conduct template button. On the left side of the page, multiple codes of conduct are listed.

Click “Contributor Covenant” and then click Review and submit to add the file to your project.

Scroll to the bottom of the page. Ensure that the radio button to commit directly to the main branch is selected and click Commit new file. Your project now contains a code of conduct.

Save the URL of CODE_OF_CONDUCT.md file in a notepad to submit later for peer review.