Create CONTRIBUTING.md
The contribution guidelines tell project participants how to contribute to the project. To add contributions guidelines, complete the following steps:

Create a new file named CONTRIBUTING.md in the root directory of the repository with the following information:

Optionally, you can review the following guides for examples of contribution guidelines and update this file.

Contributing to Legit Info, a Call for Code for Racial Justice Project
Contributing to OpenEEW
Contributing to Atom
How to contribute to Ruby on Rails
Save the URL of CONTRIBUTING.md file in a notepad to submit later for peer review.